+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## v1.0.3
+
+- Fixed `listenForToken`
+- No more full screen for `openDiscordAccount`
+- `getQrCode` return a base64 image instead of a Buffer
+- Updated api from v9 to v10
+
 ## v1.0.2
 
 - puppeteer dependencie update